% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPageXML.R
\name{getPageXML}
\alias{getPageXML}
\title{The getPageXML function}
\usage{
getPageXML(titles, file, language = "en", limit = "max")
}
\arguments{
\item{titles,}{titles of pages that need to be retrieved.}

\item{file,}{indicates where the xl file needs to be saved.}

\item{language,}{indicates the version of wikipedia to retrieve results from. Default is English wikipedia.}

\item{limit,}{indicated the maximum amount of revisions that need to be retrieved. If set to "max" the function retrieved the whole revision history.}
}
\description{
Connects to wikipedia API and retrieves an xml file with the whole revision history of an specific page.
}
\section{Warnings}{

This function relies on current version of the Wikipedia API and may need to be updated when API changes.
}
\examples{
getPageXML("Patum_de_Berga", file = 'Patum_de_Berga.xml')
}
\keyword{API,}
\keyword{Wikipedia}
\keyword{xml}

