% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRawDataPath.R
\name{getRawDataPath}
\alias{getRawDataPath}
\title{The getRawDataPath function}
\usage{
getRawDataPath(article.name)
}
\arguments{
\item{article.name,}{article name for which you want to retrieve raw data.}
}
\description{
Returns the system path of the raw data included in this package for testing purposes.
This raw data includes last 1000 revisions of
}
\examples{
path <- getRawDataPath("Vic")
}
\keyword{LCS}
\keyword{differences,}
\keyword{processing,}
\keyword{text}

