% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRevisionHistory.R
\name{getRevisionHistory}
\alias{getRevisionHistory}
\title{The getRevisionHistory function}
\usage{
getRevisionHistory(page.title, file, granularity = c("sentence", "word"))
}
\arguments{
\item{page.title,}{name of the wikipedia page to be parsed. If name includes spaces, article.name needs to include "_".}

\item{file,}{xml file containing wikipedia revision history.}

\item{granularity,}{indicated whether to look for differences in either words or sentences. If sentences are chosen, the change in only one word will look as if the whole sentence was modified.}
}
\description{
Returns wikiRevisionHistory R6-class for an specific wikipedia page, after having parsed an xml wikipedia extraction.
}
\section{Warnings}{

This function uses getDiffTexts (which uses LCS). If granularity = "word" is chosen, computation time ofr long wikipedia articles can be high. The use of "sentence" is recommended in these cases.
}
\examples{
getPageXML("Patum_de_Berga", file = 'Patum_de_Berga.xml')
Patum <- getRevisionHistory("Patum de Berga", file = 'Patum_de_Berga.xml')
Patum.words <- getRevisionHistory("Patum de Berga", file = 'Patum_de_Berga.xml', granularity = "word")
}
\keyword{R6,}
\keyword{Wikipedia}
\keyword{history}
\keyword{parser,}
\keyword{revision}
\keyword{xml}

