% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotwikiRevisionHistory.R
\name{plot.wikiRevisionHistory}
\alias{plot.wikiRevisionHistory}
\title{The plot.wikiRevisionHistory function}
\usage{
\method{plot}{wikiRevisionHistory}(x, y, fill.by = c("name", "id"),
  x.axis = c("time", "equal"), ...)
}
\arguments{
\item{x,}{name of the wikipedia page to be parsed. If name includes spaces, article.name needs to include "_".}

\item{y,}{.}

\item{fill.by,}{.}

\item{x.axis,}{.}
}
\description{
plots history flow for article revisions contained in wikiRevisionHistory R6-class.
}
\examples{
getPageXML("Patum_de_Berga", file = 'Patum_de_Berga.xml')
Patum <- getRevisionHistory("Patum de Berga", file = 'Patum_de_Berga.xml')
}
\keyword{R6,}
\keyword{Wikipedia}
\keyword{history}
\keyword{plot,}
\keyword{revision}

