% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wikiRevisionHistory.R
\docType{data}
\name{wikiRevisionHistory}
\alias{wikiRevisionHistory}
\title{The wikiRevisionHistory function}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0fc76fbc> 
 - attr(*, "name")= chr "wikiRevisionHistory_generator"
}}
\usage{
wikiRevisionHistory
}
\description{
R6 class that parses a wikipedia xml extraction and retrieves revision history for a specific wikipedia page.
}
\keyword{datasets}

